Forced unit testing singletons to be put into one compilation unit.
Added UNIT_TEST_HOOKS macro which defines all of the singleton accessor
functions needed by the unit test library. This way they all live in a
single compilation unit, which means all of the tests are a)
consolidated into a single test manager object and b) not repeated
since they're stored in a std::map. Additionally, the WindowsConsole
abstraction should only ever have the one copy.

Inlined and made static all the functions which needed it.